+++ ext/gd/libgd/gd_interpolation.c
@@ -2137,8 +2137,7 @@ gdImagePtr gdImageRotateInterpolated(const gdImagePtr src, const float angle, in
 	/* round to two decimals and keep the 100x multiplication to use it in the common square angles
 	   case later. Keep the two decimal precisions so smaller rotation steps can be done, useful for
 	   slow animations. */
-	const int angle_rounded = fmod((int) floorf(angle * 100), 360 * 100);
-
+	const int angle_rounded = (int) fmod((int) floorf(angle * 100), 360 * 100);
 	if (bgcolor < 0) {
 		return NULL;
 	}
